input-update
Now input checks if user typed anything, if not, it asks to type something, if yes, it continues with comparing the input with api, if correct - displays correct, if not - displays wrong.
I had to give card-information the id=_input to be able to query that element and replace Player 2, guess with information that user needs to type something.